+++ src/components/icons/index.ts
@@ -1,0 +1,31 @@
+import IC_CIRCLE_USER_ROUND from "./circle-user-round.svg";
+import IC_NOTEBOOK_PEN from "./notebook-pen.svg";
+import IC_PIGGY_BANK from "./piggy-bank.svg";
+import IC_WIND from "./wind.svg";
+import IC_CHART_CANDLESTICK from "./chart-candlestick.svg";
+import IC_TARGET from "./target.svg";
+import IC_UNDO_2 from "./undo-2.svg";
+import IC_STOP from "./stop.svg";
+import IC_PHONE from "./phone.svg";
+import IC_SHIELD_CHECK from "./shield-check.svg";
+import IC_BELL_RING from "./bell-ring.svg";
+import IC_CHEVRON from "./chevron.svg";
+import IC_COMPLETE from "./complete.svg";
+
+export const IconSvgs = {
+  IC_CIRCLE_USER_ROUND,
+  IC_NOTEBOOK_PEN,
+  IC_PIGGY_BANK,
+  IC_WIND,
+  IC_CHART_CANDLESTICK,
+  IC_TARGET,
+  IC_UNDO_2,
+  IC_STOP,
+  IC_PHONE,
+  IC_SHIELD_CHECK,
+  IC_BELL_RING,
+  IC_CHEVRON,
+  IC_COMPLETE,
+};
+
+export type IconSvgName = keyof typeof IconSvgs;